+++ packages/components/src/components/divider/divider-utils.ts
@@ -1,4 +1,4 @@
-export const DIVIDER_COLORS = ['contrast-low', 'contrast-medium', 'contrast-high'] as const;
+export const DIVIDER_COLORS = ['contrast-lower', 'contrast-low', 'contrast-medium', 'contrast-high'] as const;
 export type DividerColor = (typeof DIVIDER_COLORS)[number];
 
 export const DIVIDER_DIRECTIONS = ['vertical', 'horizontal'] as const;