+++ packages/components/CHANGELOG.md
@@ -93,6 +93,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - ✅Switch
 - ✅Textarea
 
+### Added
+
+- `Divider`: value `contrast-lower` for prop `color`
+
 ### Changed
 
 - In general all headings use regular as font-weight